fix: refactor nested mobile header
