+++ client-app/shared/layout/components/index.ts
@@ -1,6 +1,6 @@
 export { default as BackButtonInHeader } from "./back-button-in-header/back-button-in-header.vue";
 export { default as LoginFormSection } from "./login-form-section/login-form-section.vue";
-export { default as NestedMobileHeader } from "./nested-mobile-header/nested-mobile-header";
+export { default as NestedMobileHeader } from "./nested-mobile-header/nested-mobile-header.vue";
 export { default as SearchBar } from "./search-bar/search-bar.vue";
 export { default as SecureLayout } from "./secure-layout/secure-layout.vue";
 export { default as TwoColumn } from "./two-column.vue";
